+++ packages/clerk-js/src/ui/contexts/ClerkUIComponentsContext.tsx
@@ -1,7 +1,13 @@
 import { deprecatedObjectProperty } from '@clerk/shared/deprecated';
 import { useClerk } from '@clerk/shared/react';
 import { snakeToCamel } from '@clerk/shared/underscore';
-import type { HandleOAuthCallbackParams, OrganizationResource, UserResource } from '@clerk/types';
+import type {
+  HandleOAuthCallbackParams,
+  OrganizationResource,
+  UserButtonProps,
+  UserResource,
+  WaitlistProps,
+} from '@clerk/types';
 import React, { useCallback, useMemo } from 'react';
 
 import { SIGN_IN_INITIAL_VALUE_KEYS, SIGN_UP_INITIAL_VALUE_KEYS } from '../../core/constants';
@@ -13,7 +19,8 @@ import type { NavbarRoute } from '../elements';
 import type { ParsedQueryString } from '../router';
 import { useRouter } from '../router';
 import type {
-  AvailableComponentCtx,
+  AvailableComponentName,
+  AvailableComponentProps,
   CreateOrganizationCtx,
   GoogleOneTapCtx,
   OrganizationListCtx,
@@ -24,6 +31,7 @@ import type {
   UserButtonCtx,
   UserProfileCtx,
   UserVerificationCtx,
+  WaitlistCtx,
 } from '../types';
 import type { CustomPageContent } from '../utils';
 import {
@@ -34,7 +42,72 @@ import {
 
 const populateParamFromObject = createDynamicParamParser({ regex: /:(\w+)/ });
 
-export const ComponentContext = React.createContext<AvailableComponentCtx | null>(null);
+export function ComponentContextProvider({
+  componentName,
+  props,
+  children,
+}: {
+  componentName: AvailableComponentName;
+  props: AvailableComponentProps;
+  children: React.ReactNode;
+}) {
+  switch (componentName) {
+    case 'SignIn':
+      return <SignInContext.Provider value={{ componentName, ...props }}>{children}</SignInContext.Provider>;
+    case 'SignUp':
+      return <SignUpContext.Provider value={{ componentName, ...props }}>{children}</SignUpContext.Provider>;
+    case 'UserProfile':
+      return <UserProfileContext.Provider value={{ componentName, ...props }}>{children}</UserProfileContext.Provider>;
+    case 'UserVerification':
+      return (
+        <UserVerificationContext.Provider value={{ componentName, ...props }}>
+          {children}
+        </UserVerificationContext.Provider>
+      );
+    case 'UserButton':
+      return (
+        <UserButtonContext.Provider value={{ componentName, ...(props as UserButtonProps) }}>
+          {children}
+        </UserButtonContext.Provider>
+      );
+    case 'OrganizationSwitcher':
+      return (
+        <OrganizationSwitcherContext.Provider value={{ componentName, ...props }}>
+          {children}
+        </OrganizationSwitcherContext.Provider>
+      );
+    case 'OrganizationList':
+      return (
+        <OrganizationListContext.Provider value={{ componentName, ...props }}>
+          {children}
+        </OrganizationListContext.Provider>
+      );
+    case 'OrganizationProfile':
+      return (
+        <OrganizationProfileContext.Provider value={{ componentName, ...props }}>
+          {children}
+        </OrganizationProfileContext.Provider>
+      );
+    case 'CreateOrganization':
+      return (
+        <CreateOrganizationContext.Provider value={{ componentName, ...props }}>
+          {children}
+        </CreateOrganizationContext.Provider>
+      );
+    case 'GoogleOneTap':
+      return (
+        <GoogleOneTapContext.Provider value={{ componentName, ...props }}>{children}</GoogleOneTapContext.Provider>
+      );
+    case 'Waitlist':
+      return (
+        <WaitlistContext.Provider value={{ componentName, ...(props as WaitlistProps) }}>
+          {children}
+        </WaitlistContext.Provider>
+      );
+    default:
+      throw new Error(`Unknown component context: ${componentName}`);
+  }
+}
 
 const getInitialValuesFromQueryParams = (queryString: string, params: string[]) => {
   const props: Record<string, string> = {};
@@ -57,10 +130,13 @@ export type SignUpContextType = SignUpCtx & {
   authQueryString: string | null;
   afterSignUpUrl: string;
   afterSignInUrl: string;
+  waitlistUrl: string;
 };
 
+export const SignUpContext = React.createContext<SignUpCtx | null>(null);
+
 export const useSignUpContext = (): SignUpContextType => {
-  const { componentName, ...ctx } = (React.useContext(ComponentContext) || {}) as SignUpCtx;
+  const context = React.useContext(SignUpContext);
   const { navigate } = useRouter();
   const { displayConfig } = useEnvironment();
   const { queryParams, queryString } = useRouter();
@@ -72,6 +148,12 @@ export const useSignUpContext = (): SignUpContextType => {
     [],
   );
 
+  if (!context || context.componentName !== 'SignUp') {
+    throw new Error('Clerk: useSignUpContext called outside of the mounted SignUp component.');
+  }
+
+  const { componentName, ...ctx } = context;
+
   const redirectUrls = new RedirectUrls(
     options,
     {
@@ -81,10 +163,6 @@ export const useSignUpContext = (): SignUpContextType => {
     },
     queryParams,
   );
-
-  if (componentName !== 'SignUp') {
-    throw new Error('Clerk: useSignUpContext called outside of the mounted SignUp component.');
-  }
 
   const afterSignUpUrl = clerk.buildUrlWithAuth(redirectUrls.getAfterSignUpUrl());
   const afterSignInUrl = clerk.buildUrlWithAuth(redirectUrls.getAfterSignInUrl());
@@ -96,10 +174,12 @@ export const useSignUpContext = (): SignUpContextType => {
   // from the `path` prop instead, when the routing is set to 'path'.
   let signUpUrl = (ctx.routing === 'path' && ctx.path) || options.signUpUrl || displayConfig.signUpUrl;
   let signInUrl = ctx.signInUrl || options.signInUrl || displayConfig.signInUrl;
+  let waitlistUrl = ctx.waitlistUrl || options.waitlistUrl || displayConfig.waitlistUrl;
 
   const preservedParams = redirectUrls.getPreservedSearchParams();
   signInUrl = buildURL({ base: signInUrl, hashSearchParams: [queryParams, preservedParams] }, { stringify: true });
   signUpUrl = buildURL({ base: signUpUrl, hashSearchParams: [queryParams, preservedParams] }, { stringify: true });
+  waitlistUrl = buildURL({ base: waitlistUrl, hashSearchParams: [queryParams, preservedParams] }, { stringify: true });
 
   // TODO: Avoid building this url again to remove duplicate code. Get it from window.Clerk instead.
   const secondFactorUrl = buildURL({ base: signInUrl, hashPath: '/factor-two' }, { stringify: true });
@@ -109,6 +189,7 @@ export const useSignUpContext = (): SignUpContextType => {
     componentName,
     signInUrl,
     signUpUrl,
+    waitlistUrl,
     secondFactorUrl,
     afterSignUpUrl,
     afterSignInUrl,
@@ -129,15 +210,24 @@ export type SignInContextType = SignInCtx & {
   afterSignUpUrl: string;
   afterSignInUrl: string;
   transferable: boolean;
+  waitlistUrl: string;
 };
 
+export const SignInContext = React.createContext<SignInCtx | null>(null);
+
 export const useSignInContext = (): SignInContextType => {
-  const { componentName, ...ctx } = (React.useContext(ComponentContext) || {}) as SignInCtx;
+  const context = React.useContext(SignInContext);
   const { navigate } = useRouter();
   const { displayConfig } = useEnvironment();
   const { queryParams, queryString } = useRouter();
   const options = useOptions();
   const clerk = useClerk();
+
+  if (context === null || context.componentName !== 'SignIn') {
+    throw new Error(`Clerk: useSignInContext called outside of the mounted SignIn component.`);
+  }
+
+  const { componentName, ...ctx } = context;
 
   const initialValuesFromQueryParams = useMemo(
     () => getInitialValuesFromQueryParams(queryString, SIGN_IN_INITIAL_VALUE_KEYS),
@@ -154,10 +244,6 @@ export const useSignInContext = (): SignInContextType => {
     queryParams,
   );
 
-  if (componentName !== 'SignIn') {
-    throw new Error('Clerk: useSignInContext called outside of the mounted SignIn component.');
-  }
-
   const afterSignInUrl = clerk.buildUrlWithAuth(redirectUrls.getAfterSignInUrl());
   const afterSignUpUrl = clerk.buildUrlWithAuth(redirectUrls.getAfterSignUpUrl());
 
@@ -168,10 +254,12 @@ export const useSignInContext = (): SignInContextType => {
   // from the `path` prop instead, when the routing is set to 'path'.
   let signInUrl = (ctx.routing === 'path' && ctx.path) || options.signInUrl || displayConfig.signInUrl;
   let signUpUrl = ctx.signUpUrl || options.signUpUrl || displayConfig.signUpUrl;
+  let waitlistUrl = ctx.waitlistUrl || options.waitlistUrl || displayConfig.waitlistUrl;
 
   const preservedParams = redirectUrls.getPreservedSearchParams();
   signInUrl = buildURL({ base: signInUrl, hashSearchParams: [queryParams, preservedParams] }, { stringify: true });
   signUpUrl = buildURL({ base: signUpUrl, hashSearchParams: [queryParams, preservedParams] }, { stringify: true });
+  waitlistUrl = buildURL({ base: waitlistUrl, hashSearchParams: [queryParams, preservedParams] }, { stringify: true });
 
   const signUpContinueUrl = buildURL({ base: signUpUrl, hashPath: '/continue' }, { stringify: true });
 
@@ -181,6 +269,7 @@ export const useSignInContext = (): SignInContextType => {
     componentName,
     signUpUrl,
     signInUrl,
+    waitlistUrl,
     afterSignInUrl,
     afterSignUpUrl,
     navigateAfterSignIn,
@@ -225,14 +314,18 @@ export type UserProfileContextType = UserProfileCtx & {
   pages: PagesType;
 };
 
+export const UserProfileContext = React.createContext<UserProfileCtx | null>(null);
+
 export const useUserProfileContext = (): UserProfileContextType => {
-  const { componentName, customPages, ...ctx } = (React.useContext(ComponentContext) || {}) as UserProfileCtx;
+  const context = React.useContext(UserProfileContext);
   const { queryParams } = useRouter();
   const clerk = useClerk();
 
-  if (componentName !== 'UserProfile') {
+  if (!context || context.componentName !== 'UserProfile') {
     throw new Error('Clerk: useUserProfileContext called outside of the mounted UserProfile component.');
   }
+
+  const { componentName, customPages, ...ctx } = context;
 
   const pages = useMemo(() => {
     return createUserProfileCustomPages(customPages || [], clerk);
@@ -249,12 +342,16 @@ export const useUserProfileContext = (): UserProfileContextType => {
 
 export type UserVerificationContextType = UserVerificationCtx;
 
-export const useUserVerification = (): UserVerificationContextType => {
-  const { componentName, ...ctx } = (React.useContext(ComponentContext) || {}) as UserVerificationCtx;
+export const UserVerificationContext = React.createContext<UserVerificationCtx | null>(null);
 
-  if (componentName !== 'UserVerification') {
+export const useUserVerification = (): UserVerificationContextType => {
+  const context = React.useContext(UserVerificationContext);
+
+  if (!context || context.componentName !== 'UserVerification') {
     throw new Error('Clerk: useUserVerificationContext called outside of the mounted UserVerification component.');
   }
+
+  const { componentName, ...ctx } = context;
 
   return {
     ...ctx,
@@ -262,16 +359,20 @@ export const useUserVerification = (): UserVerificationContextType => {
   };
 };
 
+export const UserButtonContext = React.createContext<UserButtonCtx | null>(null);
+
 export const useUserButtonContext = () => {
-  const { componentName, customMenuItems, ...ctx } = (React.useContext(ComponentContext) || {}) as UserButtonCtx;
+  const context = React.useContext(UserButtonContext);
   const clerk = useClerk();
   const { navigate } = useRouter();
   const { displayConfig } = useEnvironment();
   const options = useOptions();
 
-  if (componentName !== 'UserButton') {
+  if (!context || context.componentName !== 'UserButton') {
     throw new Error('Clerk: useUserButtonContext called outside of the mounted UserButton component.');
   }
+
+  const { componentName, customMenuItems, ...ctx } = context;
 
   const signInUrl = ctx.signInUrl || options.signInUrl || displayConfig.signInUrl;
   const userProfileUrl = ctx.userProfileUrl || displayConfig.userProfileUrl;
@@ -319,14 +420,18 @@ export const useUserButtonContext = () => {
   };
 };
 
+export const OrganizationSwitcherContext = React.createContext<OrganizationSwitcherCtx | null>(null);
+
 export const useOrganizationSwitcherContext = () => {
-  const { componentName, ...ctx } = (React.useContext(ComponentContext) || {}) as OrganizationSwitcherCtx;
+  const context = React.useContext(OrganizationSwitcherContext);
   const { navigate } = useRouter();
   const { displayConfig } = useEnvironment();
 
-  if (componentName !== 'OrganizationSwitcher') {
+  if (!context || context.componentName !== 'OrganizationSwitcher') {
     throw new Error('Clerk: useOrganizationSwitcherContext called outside OrganizationSwitcher.');
   }
+
+  const { componentName, ...ctx } = context;
 
   const afterCreateOrganizationUrl = ctx.afterCreateOrganizationUrl || displayConfig.afterCreateOrganizationUrl;
   const afterLeaveOrganizationUrl = ctx.afterLeaveOrganizationUrl || displayConfig.afterLeaveOrganizationUrl;
@@ -422,14 +527,18 @@ export const useOrganizationSwitcherContext = () => {
   };
 };
 
+export const OrganizationListContext = React.createContext<OrganizationListCtx | null>(null);
+
 export const useOrganizationListContext = () => {
-  const { componentName, ...ctx } = (React.useContext(ComponentContext) || {}) as unknown as OrganizationListCtx;
+  const context = React.useContext(OrganizationListContext);
   const { navigate } = useRouter();
   const { displayConfig } = useEnvironment();
 
-  if (componentName !== 'OrganizationList') {
+  if (!context || context.componentName !== 'OrganizationList') {
     throw new Error('Clerk: useOrganizationListContext called outside OrganizationList.');
   }
+
+  const { componentName, ...ctx } = context;
 
   const afterCreateOrganizationUrl = ctx.afterCreateOrganizationUrl || displayConfig.afterCreateOrganizationUrl;
 
@@ -508,15 +617,19 @@ export type OrganizationProfileContextType = OrganizationProfileCtx & {
   isGeneralPageRoot: boolean;
 };
 
+export const OrganizationProfileContext = React.createContext<OrganizationProfileCtx | null>(null);
+
 export const useOrganizationProfileContext = (): OrganizationProfileContextType => {
-  const { componentName, customPages, ...ctx } = (React.useContext(ComponentContext) || {}) as OrganizationProfileCtx;
+  const context = React.useContext(OrganizationProfileContext);
   const { navigate } = useRouter();
   const { displayConfig } = useEnvironment();
   const clerk = useClerk();
 
-  if (componentName !== 'OrganizationProfile') {
+  if (!context || context.componentName !== 'OrganizationProfile') {
     throw new Error('Clerk: useOrganizationProfileContext called outside OrganizationProfile.');
   }
+
+  const { componentName, customPages, ...ctx } = context;
 
   const pages = useMemo(() => createOrganizationProfileCustomPages(customPages || [], clerk), [customPages]);
 
@@ -539,14 +652,18 @@ export const useOrganizationProfileContext = (): OrganizationProfileContextType 
   };
 };
 
+export const CreateOrganizationContext = React.createContext<CreateOrganizationCtx | null>(null);
+
 export const useCreateOrganizationContext = () => {
-  const { componentName, ...ctx } = (React.useContext(ComponentContext) || {}) as CreateOrganizationCtx;
+  const context = React.useContext(CreateOrganizationContext);
   const { navigate } = useRouter();
   const { displayConfig } = useEnvironment();
 
-  if (componentName !== 'CreateOrganization') {
+  if (!context || context.componentName !== 'CreateOrganization') {
     throw new Error('Clerk: useCreateOrganizationContext called outside CreateOrganization.');
   }
+
+  const { componentName, ...ctx } = context;
 
   const navigateAfterCreateOrganization = (organization: OrganizationResource) => {
     if (typeof ctx.afterCreateOrganizationUrl === 'function') {
@@ -573,15 +690,19 @@ export const useCreateOrganizationContext = () => {
   };
 };
 
+export const GoogleOneTapContext = React.createContext<GoogleOneTapCtx | null>(null);
+
 export const useGoogleOneTapContext = () => {
-  const { componentName, ...ctx } = (React.useContext(ComponentContext) || {}) as GoogleOneTapCtx;
+  const context = React.useContext(GoogleOneTapContext);
   const options = useOptions();
   const { displayConfig } = useEnvironment();
   const { queryParams } = useRouter();
 
-  if (componentName !== 'GoogleOneTap') {
+  if (!context || context.componentName !== 'GoogleOneTap') {
     throw new Error('Clerk: useGoogleOneTapContext called outside GoogleOneTap.');
   }
+
+  const { componentName, ...ctx } = context;
 
   const generateCallbackUrls = useCallback(
     (returnBackUrl: string): HandleOAuthCallbackParams => {
@@ -654,5 +775,33 @@ export const useGoogleOneTapContext = () => {
     ...ctx,
     componentName,
     generateCallbackUrls,
+  };
+};
+
+export type WaitlistContextType = WaitlistCtx & {
+  signInUrl: string;
+  redirectUrl?: string;
+};
+
+export const WaitlistContext = React.createContext<WaitlistCtx | null>(null);
+
+export const useWaitlistContext = (): WaitlistContextType => {
+  const context = React.useContext(WaitlistContext);
+  const { displayConfig } = useEnvironment();
+  const options = useOptions();
+
+  if (!context || context.componentName !== 'Waitlist') {
+    throw new Error('Clerk: useWaitlistContext called outside Waitlist.');
+  }
+
+  const { componentName, ...ctx } = context;
+
+  let signInUrl = ctx.signInUrl || options.signInUrl || displayConfig.signInUrl;
+  signInUrl = buildURL({ base: signInUrl }, { stringify: true });
+
+  return {
+    ...ctx,
+    componentName,
+    signInUrl,
   };
 };