+++ packages/expo/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Change Log
 
+## 2.2.36
+
+### Patch Changes
+
+- Updated dependencies [[`56b3f1643cff1935deb6b68ded9a429164529cad`](https://github.com/clerk/javascript/commit/56b3f1643cff1935deb6b68ded9a429164529cad)]:
+  - @clerk/clerk-js@5.31.2
+
+## 2.2.35
+
+### Patch Changes
+
+- Updated dependencies [[`9557b557489c60fbd067906afa97a7c418f6569e`](https://github.com/clerk/javascript/commit/9557b557489c60fbd067906afa97a7c418f6569e), [`8a04ae47b8305f994b348301fd8134d5baf02943`](https://github.com/clerk/javascript/commit/8a04ae47b8305f994b348301fd8134d5baf02943)]:
+  - @clerk/clerk-js@5.31.1
+  - @clerk/clerk-react@5.15.1
+  - @clerk/shared@2.11.5
+
+## 2.2.34
+
+### Patch Changes
+
+- Updated dependencies [[`d74a6a7c0f291104c6bba722a8c432814d7b336e`](https://github.com/clerk/javascript/commit/d74a6a7c0f291104c6bba722a8c432814d7b336e), [`1a0c8fe665869e732d3c800bde0f5219fce54301`](https://github.com/clerk/javascript/commit/1a0c8fe665869e732d3c800bde0f5219fce54301), [`0800fc3f1f4e1b6a1d13f5c02557001a283af6e8`](https://github.com/clerk/javascript/commit/0800fc3f1f4e1b6a1d13f5c02557001a283af6e8)]:
+  - @clerk/clerk-js@5.31.0
+  - @clerk/clerk-react@5.15.0
+  - @clerk/types@4.30.0
+  - @clerk/shared@2.11.4
+
 ## 2.2.33
 
 ### Patch Changes